Backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
openai
sqlalchemy
psycopg2-binary
authlib
python-multipart
passlib==1.7.4
bcrypt==4.0.1
email-validator
